requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==4.4.2
et-xmlfile==1.1.0
idna==3.3
imageio==2.10.1
imageio-ffmpeg==0.4.5
lxml==4.6.4
moviepy==1.0.3
numpy==1.21.3
openpyxl==3.0.9
Pillow==8.4.0
proglog==0.1.9
python-docx==0.8.11
python-pptx==0.6.21
reportlab==3.6.2
requests==2.26.0
tqdm==4.62.3
urllib3==1.26.7
XlsxWriter==3.0.2
